Feature: floating point font size

+++ src/cornui/xml/dom_node_getter_setter.cpp
@@ -70,7 +70,7 @@ namespace cornui {
                 // todo: Use default font
                 throw std::logic_error("Font not found: " + this->computedStyle_.at("font-family"));
             }
-            size_t fontSize = std::stoi(this->computedStyle_.at("font-size"));
+            float fontSize = std::stof(this->computedStyle_.at("font-size"));
             corn::Color fontColor = corn::Color::parse(this->computedStyle_.at("font-color"));
             corn::FontVariant fontVariant = corn::FontVariant::REGULAR;
             if (this->computedStyle_.at("font-variant") == "bold") {
